requirements.txt
Django
django-photologue
django-blog-zinnia
#MySQL-Python
mysqlclient
django-recaptcha
django-admin-tools
admin-tools-zinnia
zinnia-wysiwyg-ckeditor
